refactor: move backoff logic, cleanup tests
