+++ src/substantial/src/lib.rs
@@ -2,6 +2,10 @@
 // SPDX-License-Identifier: MPL-2.0
 
 pub mod backends;
+pub mod backoff;
 pub mod converters;
 pub mod protocol;
 pub mod run;
+
+#[cfg(test)]
+mod tests;
