+++ src/typegate/engine/00_runtime.js
@@ -79,6 +79,7 @@ const Meta = {
       "op_sub_metadata_enumerate_all_children",
     ),
     runEnsureDeterminism: getOp("op_sub_run_ensure_determinism"),
+    strategyRetry: getOp("op_sub_strategy_retry"),
   },
   grpc: {
     register: getOp("op_grpc_register"),
